+++ CHANGELOG.md
@@ -35,6 +35,7 @@ All notable changes to this project will be documented in this file.
 - Reload button disabled after saving new file
 - PlaceObject tag - do not display export name twice
 - Loading nested characters when Importassets tag used
+- Hide various actions for imported tags
 
 ### Changed
 - Quick search needs minimum of 3 characters